Add algorithm for match

+++ Domain/Sources/Domain/Entities/SwipeType.swift
@@ -2,5 +2,5 @@ import Foundation
 
 public enum SwipeType {
     case left
-    case right(onMath: () -> Void)
+    case right(onMath: (Int) -> Void)
 }
